pyproject.toml
[tool.poetry]
name = "py-LabUnAL_pkg"
version = "0.0.0"
description = "LABel UNcertainty using Active Learning"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
click = "*"
tqdm = "^4.50.2"
torch = "^1.6.0"
torchvision = "^0.7.0"
scipy = "^1.5.3"
scikit-image = "0.15.0"
Pillow = "6.2.1"
opencv-python = "^4.4.0"

[tool.poetry.dev-dependencies]
pytest = "*"
sphinx_rtd_theme = "*"
sphinx = "*"
pre_commit = "*"

[tool.poetry.scripts]
py-LabUnAL_pkg = "py-LabUnAL_pkg.cli:cli"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["LabUnAL_pkg", "PIL", "active_selection", "click", "constants", "cv2", "dataloader", "image_transforms", "model", "numpy", "scipy", "skimage", "torch", "torchvision", "tqdm", "trainer", "utils"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
